refactor diaries tests: seperate diary and naver
