+++ english_diary/diaries/tests/naver/test_dictionary.py
