+++ english_diary/diaries/tests/naver/test_translate.py
